logout logics

+++ backend/src/auth/auth.service.ts
@@ -2,21 +2,22 @@ import { CreateUserDto } from '@dto/create-user.dto';
 import { GenerateEmailConfirmResponseDto } from '@dto/generate-email-confirm-response.dto';
 import { LoginRequestDto } from '@dto/login-request.dto';
 import { LoginResponseDto } from '@dto/login-response.dto';
-import {
-  Injectable,
-  UnauthorizedException,
-  ForbiddenException,
-  BadRequestException,
-} from '@nestjs/common';
+import { Injectable, BadRequestException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { EmailConfirmService } from 'src/email-confirm/email-confirm.service';
 import { RedisService } from 'src/redis/redis.service';
 import { matchPassword } from 'src/user/user-password.helper';
 import { UserService } from 'src/user/user.service';
 
+interface Payload {
+  sub: number;
+  username: string;
+}
+
 @Injectable()
 export class AuthService {
   readonly TOKEN_REFRESH_EXPIRATION = 7 * 24 * 60 * 60;
+  readonly ACCESS_TOKEN_EXPIRATION = '15m';
   readonly REFRESH_TOKEN_EXPIRATION = '30d';
 
   constructor(
@@ -30,19 +31,16 @@ export class AuthService {
     const user = await this.usersSrv.findOne(login.name);
 
     if (!(await matchPassword(user?.password, login.password))) {
-      throw new UnauthorizedException();
+      throw new BadRequestException();
     }
 
     if (!user.isMailConfirm) {
-      throw new ForbiddenException();
+      throw new BadRequestException();
     }
 
     const payload = { sub: user.id, username: user.name };
-    const accessToken = this.jwtSrv.sign(payload);
-    const refreshToken = this.jwtSrv.sign(payload, {
-      secret: process.env.JWT_REFRESH_KEY,
-      expiresIn: this.REFRESH_TOKEN_EXPIRATION,
-    });
+    const accessToken = this.generateAccessToken(payload);
+    const refreshToken = this.generateAccessToken(payload);
 
     await this.redisSrv.setRefreshToken(
       user.id,
@@ -93,17 +91,9 @@ export class AuthService {
       throw new BadRequestException();
     }
 
-    const newAccessToken = this.jwtSrv.sign({
-      sub: user.id,
-      username: user.name,
-    });
-    const newRefreshToken = this.jwtSrv.sign(
-      { sub: user.id, username: user.name },
-      {
-        secret: process.env.JWT_REFRESH_SECRET,
-        expiresIn: this.REFRESH_TOKEN_EXPIRATION,
-      },
-    );
+    const payload = { sub: user.id, username: user.name };
+    const newAccessToken = this.generateAccessToken(payload);
+    const newRefreshToken = this.generateRefreshToken(payload);
 
     await this.redisSrv.setRefreshToken(
       user.id,
@@ -119,7 +109,25 @@ export class AuthService {
   }
 
   async logout(userId: number) {
-    await this.redisSrv.deleteRefreshToken(userId);
-    await this.usersSrv.invalidateRefreshToken(userId);
+    if (await this.redisSrv.getUserRefreshToken(userId)) {
+      await this.redisSrv.deleteRefreshToken(userId);
+      await this.usersSrv.invalidateRefreshToken(userId);
+    } else {
+      throw new BadRequestException();
+    }
+  }
+
+  private generateAccessToken(payload: Payload): string {
+    return this.jwtSrv.sign(payload, {
+      secret: process.env.JWT_ACCESS_KEY,
+      expiresIn: this.ACCESS_TOKEN_EXPIRATION,
+    });
+  }
+
+  private generateRefreshToken(payload: Payload): string {
+    return this.jwtSrv.sign(payload, {
+      secret: process.env.JWT_REFRESH_KEY,
+      expiresIn: this.REFRESH_TOKEN_EXPIRATION,
+    });
   }
 }
